+++ large_image/tilesource/base.py
@@ -1603,7 +1603,8 @@ class TileSource:
         :returns: an integer frame number.
         """
         frame = int(frame or 0)
-        if (hasattr(self, 'style') and 'bands' in self.style and
+        if (not getattr(self, '_skipStyle', None) and
+                hasattr(self, 'style') and 'bands' in self.style and
                 len(self.style['bands']) and
                 all(entry.get('frame') is not None for entry in self.style['bands'])):
             frame = int(self.style['bands'][0]['frame'])