+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Show and edit yaml and json files using codemirror ([#969](../../pull/969), [#971](../../pull/971))
 - Show configured item lists even if there are no large images ([#972](../../pull/972))
 - Add metadata and annotation metadata search modes to Girder ([#974](../../pull/974))
+- Add the ability to show annotation metadata in item annotation lists ([#977](../../pull/977))
 
 ## 1.17.0