+++ CHANGELOG.md
@@ -9,7 +9,8 @@
 - Add metadata and annotation metadata search modes to Girder ([#974](../../pull/974))
 - Add the ability to show annotation metadata in item annotation lists ([#977](../../pull/977))
 - Support ETAG in annotation rest responses for better browser caching ([#978](../../pull/978))
-- Thumbnail maintenance endpoints ([#979](../../pull/979))
+- Thumbnail maintenance endpoints ([#980](../../pull/980))
+- Handle lif file extensions ([#981](../../pull/981))
 
 ## 1.17.0